Refactor differences into their own namespace
